+++ src/components/IntegrationsView.tsx
@@ -18,7 +18,6 @@ import { useSystemAudioPermission } from "../hooks/useSystemAudioPermission";
 import { canManageSystemAudioInApp } from "../utils/systemAudioAccess";
 import ApiKeysSection from "./ApiKeysSection";
 import CliIntegrationCard from "./CliIntegrationCard";
-import McpIntegrationCard from "./McpIntegrationCard";
 import googleCalendarIcon from "../assets/icons/google-calendar.svg";
 
 const API_DOCS_URL = "https://docs.openwhispr.com/api/overview";
@@ -248,11 +247,6 @@ export default function IntegrationsView({ isPaid, onUpgrade }: IntegrationsView
             </div>
           </SettingsPanelRow>
         </SettingsPanel>
-      </div>
-
-      <div>
-        <SectionLabel>{t("integrations.sections.mcp")}</SectionLabel>
-        <McpIntegrationCard isPaid={isPaid} onUpgrade={onUpgrade} />
       </div>
 
       <div>